+++ scrybe_frontend/src/Auth.jsx
@@ -1,10 +1,8 @@
-import { Navigate, Outlet, useLocation } from "react-router-dom";
+import { Navigate, Outlet } from "react-router-dom";
 import { localStorageUser } from "./helpers/localStorageUser";
 
 export function RequireToken() {
-  //let auth =   localStorageUser() {name: "user", token: "token"}};
-  let auth = { name: "user", token: "token" };
-  let location = useLocation();
+  let auth = localStorageUser();
   if (!auth) {
     return <Navigate to="/login" replace />;
   }
