+++ scrybe_frontend/src/pages/UploadedRecordings/TableData/tabledata.module.scss
