+++ scrybe_frontend/src/pages/UploadedRecordings/uploadedRecordings.module.scss
